A deep-learning-based SIoV framework in vehicle detection and counting system for Intelligent traffic management
The automatic detection and counting of vehicles
can greatly contribute to improving traffic control and
management. This is done by utilizing a deep neural network
and classifying labeled data. The proposed model has three
stages: first, to separate and count the vehicles; second, to
determine whether the vehicle is slow or fast; and third, to make
a decision about each vehicle. In the first step, vehicle images are
analyzed using the YOLOv3 method, and the vehicles detected
are counted. In the next stage, the labeled data is classified into
two classes, slow and fast, using a recurrent neural network.
Finally, considering the road type, number of cars, and the class
of each car, a decision is made about the existing traffic
situation. The proposed method was evaluated on two datasets,
GRAM and HighwayII, and the experimental results show an
accuracy of about 98%, which is better than other existing
approaches.